requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.15.0
numpy==1.26.4
pillow==10.4.0
mlflow==2.15.1
pytest==8.3.2
python-multipart